UTF-8 charset and FidgetExpress update
