+++ public_html/connect.php
@@ -10,5 +10,6 @@ $password = $url['pass'];
 $db = substr($url['path'], 1);
 
 $conn = new mysqli($server, $username, $password, $db);
+$conn->set_charset('utf8');
 
 ?>
