+++ public_html/header.php
@@ -1,5 +1,6 @@
 <head>
-	<title>W3bShopAtt4ckzz</title>
-	<link rel='icon' type='image/png' href='img/icon.png'>
+	<title>Fidget Express</title>
+	<meta charset="utf-8">
+	<link rel='icon' type='image/png' href='img/fidget.png'>
 	<link rel="stylesheet" href="css/index.css" />
 </head>
