+++ public_html/navigationBar.php
@@ -1,7 +1,15 @@
 <?php $site = basename($_SERVER['PHP_SELF']); ?>
 <div class="navigationBar">
-	<a href="store.php"<?php if($site == "store.php"){ echo 'class="active"'; } ?>>Produkter</a>
-	<a href="logout.php">Logga ut</a>
-	<a href="profile.php"<?php if($site == "profile.php"){ echo 'class="active"'; } ?>>Min profil (<?= $_SESSION['username'] ?>)</a>
-	<a href="cart.php"<?php if($site == "cart.php"){ echo 'class="active"'; } ?>>Kundvagn</a>
+	<a href="store.php"<?= $site == "store.php" ? 'class="active"' : '' ?>>
+		Produkter
+	</a>
+	<a href="logout.php">
+		Logga ut
+	</a>
+	<a href="profile.php"<?= $site == "profile.php" ? 'class="active"' : '' ?>>
+		Min profil (<?= $_SESSION['username']?>)
+	</a>
+	<a href="cart.php"<?= $site == "cart.php" ? 'class="active"' : '' ?>>
+		Kundvagn
+	</a>
 </div>
